requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coverage==5.3.1
Django==3.1.3
django-debug-toolbar==3.2
djangorestframework==3.12.2
idna==2.10
python-slugify==4.0.1
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
